Rename async to async_mode
Fixes W606 'async' and 'await' are reserved keywords starting with Python 3.7

This continues to accept `--async` as an argument and `async = 1` in
config files.  Therefore this commit should not be breaking.
