+++ pylama/config.py
@@ -131,7 +131,7 @@ PARSER.add_argument(
     "--hook", action="store_true", help="Install Git (Mercurial) hook.")
 
 PARSER.add_argument(
-    "--async", action="store_true",
+    "--async", action="store_true", dest='async_mode',
     help="Enable async mode. Useful for checking a lot of files. "
     "Unsupported with pylint.")
 
@@ -172,6 +172,8 @@ def parse_options(args=None, config=True, rootdir=CURDIR, **overrides):  # noqa
     if config:
         cfg = get_config(str(options.options), rootdir=rootdir)
         for opt, val in cfg.default.items():
+            if opt == 'async':
+                opt = 'async_mode'
             LOGGER.info('Find option %s (%s)', opt, val)
             passed_value = getattr(options, opt, _Default())
             if isinstance(passed_value, _Default):
@@ -203,9 +205,9 @@ def parse_options(args=None, config=True, rootdir=CURDIR, **overrides):  # noqa
         if isinstance(value, _Default):
             setattr(options, name, process_value(name, value.value))
 
-    if options.async and 'pylint' in options.linters:
+    if options.async_mode and 'pylint' in options.linters:
         LOGGER.warning('Can\'t parse code asynchronously with pylint enabled.')
-        options.async = False
+        options.async_mode = False
 
     return options
