+++ test_pylama.py
@@ -1,6 +1,6 @@
 import os.path as op
 
-from pylama.async import check_async
+from pylama.async_mode import check_async
 from pylama.config import parse_options, get_config
 from pylama.core import filter_errors, parse_modeline, prepare_params, run
 from pylama.errors import Error, remove_duplicates
